requirements.txt
asgiref==3.8.1
black==25.1.0
certifi==2025.4.26
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.1.8
distlib==0.3.9
Django==5.2.1
djangorestframework==3.16.0
filelock==3.18.0
flake8==7.2.0
identify==2.6.10
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pre_commit==4.2.0
pycodestyle==2.13.0
pyflakes==3.3.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
ruff==0.11.8
sqlparse==0.5.3
urllib3==2.4.0
virtualenv==20.31.1
